feat: Merge PaginatedOperator to SimpleHttpOperator
