+++ airflow/providers/http/operators/http.py
@@ -19,15 +19,16 @@ from __future__ import annotations
 
 import base64
 import pickle
-from typing import TYPE_CHECKING, Any, Callable, Generic, Sequence, TypeVar, Union, cast
+from typing import TYPE_CHECKING, Any, Callable, Sequence
 
-from requests import Response as RequestResponse
+from requests import Response
 
 from airflow.configuration import conf
 from airflow.exceptions import AirflowException
 from airflow.hooks.base import BaseHook
 from airflow.models import BaseOperator
 from airflow.providers.http.triggers.http import HttpTrigger
+from airflow.utils.helpers import merge_dicts
 
 if TYPE_CHECKING:
     from requests.auth import AuthBase
@@ -36,13 +37,7 @@ if TYPE_CHECKING:
     from airflow.utils.context import Context
 
 
-# By default, SimpleHttpOperator pass a request.Response object as parameter to
-# injected functions, and return a str object (from request.Response.text) as result.
-Response = TypeVar("Response", bound=Union[RequestResponse, Any])
-ResponseText = TypeVar("ResponseText", bound=Union[str, Any])
-
-
-class SimpleHttpOperator(BaseOperator, Generic[Response, ResponseText]):
+class SimpleHttpOperator(BaseOperator):
     """
     Calls an endpoint on an HTTP system to execute an action.
 
@@ -57,14 +52,26 @@ class SimpleHttpOperator(BaseOperator, Generic[Response, ResponseText]):
     :param data: The data to pass. POST-data in POST/PUT and params
         in the URL for a GET request. (templated)
     :param headers: The HTTP headers to be added to the GET request
+    :param pagination_function: A callable that generates the parameters used to call the API again.
+        Typically used when the API is paginated and returns for e.g a cursor, a 'next page id', or
+        a 'next page URL'. When provided, the Operator will call the API repeatedly until this callable
+        returns None. Also, the result of the Operator will become by default a list of Response.text
+        objects (instead of a single response object). Same with the other injected functions (like
+        response_check, response_filter, ...) which will also receive a list of Response object. This
+        function should return a dict of parameters (`endpoint`, `data`, `headers`, `extra_options`),
+        which will be merged and override the one used in the initial API call.
     :param response_check: A check against the 'requests' response object.
         The callable takes the response object as the first positional argument
         and optionally any number of keyword arguments available in the context dictionary.
-        It should return True for 'pass' and False otherwise.
+        It should return True for 'pass' and False otherwise. If a pagination_function
+        is provided, this function will receive a list of response objects instead of a
+        single response object.
     :param response_filter: A function allowing you to manipulate the response
         text. e.g response_filter=lambda response: json.loads(response.text).
         The callable takes the response object as the first positional argument
         and optionally any number of keyword arguments available in the context dictionary.
+        If a pagination_function is provided, this function will receive a list of response
+        object instead of a single response object.
     :param extra_options: Extra options for the 'requests' library, see the
         'requests' documentation (options to modify timeout, ssl, etc.)
     :param log_response: Log the response (default: False)
@@ -94,6 +101,7 @@ class SimpleHttpOperator(BaseOperator, Generic[Response, ResponseText]):
         method: str = "POST",
         data: Any = None,
         headers: dict[str, str] | None = None,
+        pagination_function: Callable[..., Any] | None = None,
         response_check: Callable[..., bool] | None = None,
         response_filter: Callable[..., Any] | None = None,
         extra_options: dict[str, Any] | None = None,
@@ -113,6 +121,7 @@ class SimpleHttpOperator(BaseOperator, Generic[Response, ResponseText]):
         self.endpoint = endpoint
         self.headers = headers or {}
         self.data = data or {}
+        self.pagination_function = pagination_function
         self.response_check = response_check
         self.response_filter = response_filter
         self.extra_options = extra_options or {}
@@ -152,7 +161,21 @@ class SimpleHttpOperator(BaseOperator, Generic[Response, ResponseText]):
     def execute_sync(self, context: Context) -> Any:
         self.log.info("Calling HTTP method")
         response = self.hook.run(self.endpoint, self.data, self.headers, self.extra_options)
+        response = self.paginate_sync(first_response=response)
         return self.process_response(context=context, response=response)
+
+    def paginate_sync(self, first_response: Response) -> Response | list[Response]:
+        if not self.pagination_function:
+            return first_response
+
+        all_responses = [first_response]
+        while True:
+            next_page_params = self.pagination_function(first_response)
+            if not next_page_params:
+                break
+            response = self.hook.run(**self._merge_next_page_parameters(next_page_params))
+            all_responses.append(response)
+        return all_responses
 
     def execute_async(self, context: Context) -> None:
         self.defer(
@@ -168,12 +191,14 @@ class SimpleHttpOperator(BaseOperator, Generic[Response, ResponseText]):
             method_name="execute_complete",
         )
 
-    def process_response(self, context: Context, response: Response) -> Any:
+    def process_response(self, context: Context, response: Response | list[Response]) -> Any:
         """Process the response."""
         from airflow.utils.operator_helpers import determine_kwargs
 
+        make_default_response: Callable = self._default_response_maker(response=response)
+
         if self.log_response:
-            self.log.info(self.default_response_maker(response))
+            self.log.info(make_default_response())
         if self.response_check:
             kwargs = determine_kwargs(self.response_check, [response], context)
             if not self.response_check(response, **kwargs):
@@ -181,13 +206,24 @@ class SimpleHttpOperator(BaseOperator, Generic[Response, ResponseText]):
         if self.response_filter:
             kwargs = determine_kwargs(self.response_filter, [response], context)
             return self.response_filter(response, **kwargs)
-        return self.default_response_maker(response)
+        return make_default_response()
 
     @staticmethod
-    def default_response_maker(response: Response) -> ResponseText:
-        return cast(ResponseText, response.text)
+    def _default_response_maker(response: Response | list[Response]) -> Callable:
+        """Create a default response maker function based on the type of response.
+        :param response: The response object or list of response objects.
+        :return: A function that returns response text(s).
+        """
+        if isinstance(response, Response):
+            response_object = response  # Makes mypy happy
+            return lambda: response_object.text
 
-    def execute_complete(self, context: Context, event: dict):
+        response_list: list[Response] = response  # Makes mypy happy
+        return lambda: [entry.text for entry in response_list]
+
+    def execute_complete(
+        self, context: Context, event: dict, paginated_responses: None | list[Response] = None
+    ):
         """
         Callback for when the trigger fires - returns immediately.
 
@@ -195,6 +231,46 @@ class SimpleHttpOperator(BaseOperator, Generic[Response, ResponseText]):
         """
         if event["status"] == "success":
             response = pickle.loads(base64.standard_b64decode(event["response"]))
+
+            self.paginate_async(context=context, response=response, previous_responses=paginated_responses)
             return self.process_response(context=context, response=response)
         else:
             raise AirflowException(f"Unexpected error in the operation: {event['message']}")
+
+    def paginate_async(
+        self, context: Context, response: Response, previous_responses: None | list[Response] = None
+    ):
+        if self.pagination_function:
+            all_responses = previous_responses or []
+            all_responses.append(response)
+
+            next_page_params = self.pagination_function(response)
+            if not next_page_params:
+                return self.process_response(context=context, response=all_responses)
+            self.defer(
+                trigger=HttpTrigger(
+                    http_conn_id=self.http_conn_id,
+                    auth_type=self.auth_type,
+                    method=self.method,
+                    **self._merge_next_page_parameters(next_page_params),
+                ),
+                method_name="execute_complete",
+                kwargs={"paginated_responses": all_responses},
+            )
+
+    def _merge_next_page_parameters(self, next_page_params: dict) -> dict:
+        """Merge initial request parameters with next page parameters.
+
+        Merge initial requests parameters with the ones for the next page, generated by
+        the pagination function. Items in the 'next_page_params' overrides those defined
+        in the previous request.
+
+        :param next_page_params: A dictionary containing the parameters for the next page.
+        :return: A dictionary containing the merged parameters.
+        """
+        return dict(
+            endpoint=next_page_params.get("endpoint") or self.endpoint,
+            data=merge_dicts(self.data, next_page_params.get("data", {})),
+            headers=merge_dicts(self.headers, next_page_params.get("headers", {})),
+            extra_options=merge_dicts(self.extra_options, next_page_params.get("extra_options", {})),
+        )
