+++ airflow/providers/http/provider.yaml
@@ -68,7 +68,6 @@ operators:
   - integration-name: Hypertext Transfer Protocol (HTTP)
     python-modules:
       - airflow.providers.http.operators.http
-      - airflow.providers.http.operators.paginated
 sensors:
   - integration-name: Hypertext Transfer Protocol (HTTP)
     python-modules:
