+++ tests/system/providers/http/example_http.py
@@ -24,7 +24,6 @@ from datetime import datetime
 
 from airflow import DAG
 from airflow.providers.http.operators.http import SimpleHttpOperator
-from airflow.providers.http.operators.paginated import PaginatedHttpOperator
 from airflow.providers.http.sensors.http import HttpSensor
 
 ENV_ID = os.environ.get("SYSTEM_TESTS_ENV_ID")
@@ -125,7 +124,7 @@ def get_next_page_cursor(response) -> dict | None:
         return dict(data={"cursor": next_cursor})
 
 
-task_get_paginated = PaginatedHttpOperator(
+task_get_paginated = SimpleHttpOperator(
     task_id="get_paginated",
     method="GET",
     endpoint="get",
